サンプルの画面実装をuiモジュールに移動
