--- .idea/runConfigurations/melos_flutter_run_practice.xml
@@ -1,6 +1,0 @@
-<component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="Flutter Run -&gt; 'practice'" type="FlutterRunConfigurationType" factoryName="Flutter">
-    <option name="filePath" value="$PROJECT_DIR$/packages/ui/practice/lib/main.dart" />
-    <method v="2" />
-  </configuration>
-</component>
