+++ .idea/runConfigurations/melos_flutter_test_practice.xml
@@ -1,0 +1,7 @@
+<!-- Generated by Melos -->
+<component name="ProjectRunConfigurationManager">
+  <configuration default="false" name="Flutter Test -&gt; 'practice'" type="FlutterTestConfigType" factoryName="Flutter Test">
+    <option name="testDir" value="$PROJECT_DIR$/packages/ui/practice/test" />
+    <method v="2" />
+  </configuration>
+</component>
