+++ packages/ui/practice/lib/practice.dart
@@ -1,0 +1,1 @@
+export 'src/practice_screen.dart';
