+++ packages/ui/practice/lib/src/practice_screen.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+class PracticeScreen extends ConsumerWidget {
+  const PracticeScreen({super.key});
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    return Scaffold(
+      appBar: AppBar(title: const Text('Bass App')),
+      body: const Center(child: Text('日本語の確認サンプル')),
+    );
+  }
+}
