+++ packages/ui/practice/pubspec_overrides.yaml
@@ -1,4 +1,4 @@
-# melos_managed_dependency_overrides: chord_progression,core,metronome,practice
+# melos_managed_dependency_overrides: chord_progression,core,metronome
 dependency_overrides:
   chord_progression:
     path: ../../features/chord_progression
@@ -6,5 +6,3 @@ dependency_overrides:
     path: ../../core
   metronome:
     path: ../../features/metronome
-  practice:
-    path: ../../ui/practice
